requirements.txt
groq
ffmpeg
python-dotenv
portaudio
speechrecognition
pydubpyaudio
elevenlabs